refactor: roving-focus
