+++ packages/core/roving-focus/src/RovingFocusGroup.ts
@@ -1,78 +1,9 @@
-import { createProvideScope } from '@oku-ui/provide'
-import type { CollectionPropsType } from '@oku-ui/collection'
-import { createCollection } from '@oku-ui/collection'
-import type { ComputedRef, Ref } from 'vue'
 import { defineComponent, h, mergeProps } from 'vue'
 import { useForwardRef } from '@oku-ui/use-composable'
 import { primitiveProps } from '@oku-ui/primitive'
-import { OkuRovingFocusGroupImpl, rovingFocusGroupImplProps } from './RovingFocusGroupImpl'
-import type { RovingFocusGroupImplElement, RovingFocusGroupImplNaviteElement, RovingFocusGroupImplProps } from './RovingFocusGroupImpl'
-import { scopedProps } from './types'
-import type { Direction, Orientation } from './utils'
-
-const GROUP_NAME = 'OkuRovingFocusGroup'
-
-export interface ItemData extends CollectionPropsType {
-  id: string
-  focusable: boolean
-  active: boolean
-}
-
-export const { CollectionItemSlot, CollectionProvider, CollectionSlot, useCollection, createCollectionScope } = createCollection<
-  HTMLSpanElement,
-  ItemData
->(GROUP_NAME, {
-  id: {
-    type: String,
-  },
-  focusable: {
-    type: Boolean,
-  },
-  active: {
-    type: Boolean,
-  },
-})
-
-const [createRovingFocusGroupProvide, createRovingFocusGroupScope] = createProvideScope(
-  GROUP_NAME,
-  [createCollectionScope],
-)
-
-// RovingFocusGroupOptions extends
-type RovingProvideValue = {
-  /**
- * The orientation of the group.
- * Mainly so arrow navigation is done accordingly (left & right vs. up & down)
- */
-  orientation?: Ref<Orientation | undefined>
-  /**
-   * The direction of navigation between items.
-   */
-  dir?: Ref<Direction | undefined>
-  /**
-   * Whether keyboard navigation should loop around
-   * @defaultValue false
-   */
-  loop?: Ref<boolean | undefined>
-
-  currentTabStopId: ComputedRef<string | null>
-  onItemFocus(tabStopId: string): void
-  onItemShiftTab(): void
-  onFocusableItemAdd(): void
-  onFocusableItemRemove(): void
-}
-
-export const [rovingFocusProvider, useRovingFocusInject]
-  = createRovingFocusGroupProvide<RovingProvideValue>(GROUP_NAME)
-
-export type RovingFocusGroupNaviteElement = RovingFocusGroupImplNaviteElement
-export type RovingFocusGroupElement = RovingFocusGroupImplElement
-
-export interface RovingFocusGroupProps extends RovingFocusGroupImplProps { }
-
-export const rovingFocusGroupProps = {
-  ...rovingFocusGroupImplProps,
-}
+import { OkuRovingFocusGroupImpl } from './RovingFocusGroupImpl'
+import type { RovingFocusGroupNaviteElement } from './props'
+import { CollectionItemSlot, CollectionProvider, CollectionSlot, GROUP_NAME, rovingFocusGroupProps, scopedProps } from './props'
 
 const rovingFocusGroup = defineComponent({
   name: GROUP_NAME,
@@ -109,12 +40,7 @@ const rovingFocusGroup = defineComponent({
 })
 
 // TODO: https://github.com/vuejs/core/pull/7444 after delete
-const OkuRovingFocusGroup = rovingFocusGroup as typeof rovingFocusGroup &
+export const OkuRovingFocusGroup = rovingFocusGroup as typeof rovingFocusGroup &
 (new () => {
   $props: RovingFocusGroupNaviteElement
 })
-
-export {
-  OkuRovingFocusGroup,
-  createRovingFocusGroupScope,
-}
