+++ packages/core/roving-focus/src/RovingFocusGroupImpl.ts
@@ -1,59 +1,11 @@
-import type { PropType } from 'vue'
 import { computed, defineComponent, h, mergeProps, reactive, ref, toRefs, watchEffect } from 'vue'
 import { reactiveOmit, useComposedRefs, useControllable, useForwardRef } from '@oku-ui/use-composable'
 
-import type { OkuElement } from '@oku-ui/primitive'
-
 import { Primitive, primitiveProps } from '@oku-ui/primitive'
 import { composeEventHandlers } from '@oku-ui/utils'
-import type { RovingFocusGroupOptions } from './utils'
-import { focusFirst, rovingFocusGroupOptionsProps } from './utils'
-import { rovingFocusProvider, useCollection } from './RovingFocusGroup'
-import { scopedProps } from './types'
-
-const ENTRY_FOCUS = 'rovingFocusGroup.onEntryFocus'
-const EVENT_OPTIONS = { bubbles: false, cancelable: true }
-
-export type RovingFocusGroupImplNaviteElement = OkuElement<'div'>
-export type RovingFocusGroupImplElement = HTMLDivElement
-
-export interface RovingFocusGroupImplProps extends RovingFocusGroupOptions {
-  currentTabStopId?: string | null
-  defaultCurrentTabStopId?: string
-}
-
-export type RovingFocusGroupImplEmits = {
-  currentTabStopIdChange: [tabStopId: string | null]
-  entryFocus: [event: Event]
-  mousedown: [event: MouseEvent]
-  focus: [event: FocusEvent]
-  blur: [event: FocusEvent]
-}
-
-export const rovingFocusGroupImplProps = {
-  props: {
-    currentTabStopId: {
-      type: String as PropType<string | null>,
-    },
-    defaultCurrentTabStopId: {
-      type: String as PropType<string>,
-      default: null,
-    },
-    ...rovingFocusGroupOptionsProps.props,
-  },
-  emits: {
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    entryFocus: (event: Event) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    currentTabStopIdChange: (tabStopId: string | null) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    mousedown: (event: MouseEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    focus: (event: FocusEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    blur: (event: FocusEvent) => true,
-  },
-}
+import { focusFirst } from './utils'
+import type { RovingFocusGroupImplNaviteElement } from './props'
+import { ENTRY_FOCUS, EVENT_OPTIONS, rovingFocusGroupImplProps, rovingFocusProvider, scopedProps, useCollection } from './props'
 
 const RovingFocusGroupImpl = defineComponent({
   name: 'OkuRovingFocusGroupImpl',
