--- packages/core/roving-focus/src/types.ts
@@ -1,8 +1,0 @@
-import { type Scope, ScopePropObject } from '@oku-ui/provide'
-
-export type ScopedPropsInterface<P> = P & { scopeOkuRovingFocusGroup?: Scope }
-export const scopedProps = {
-  scopeOkuRovingFocusGroup: {
-    ...ScopePropObject,
-  },
-}
